Split CompiledCall into compiled_call.rs to satisfy the file-size budget
The Quality Gate's file-size guard (SPEC §14.1) flagged compiled_plan.rs
at 577 lines against its 508-line grandfathered baseline (+10% ceiling
558). Move the CompiledCall site-specialization struct and its executor
into their own module instead of re-baselining: compiled_plan.rs returns
to 516 lines and the new compiled_call.rs (71 lines) is well under the
500-line budget for new files. No behavior change.
